refactored table prop types
