+++ src/js/prop-types/table-scope-prop-type.js
@@ -1,0 +1,11 @@
+import PropTypes from './prop-types';
+
+const tableScopePropType = PropTypes.oneOf([
+  'row',
+  'col',
+  'rowgroup',
+  'colgroup',
+  'auto',
+]);
+
+export default tableScopePropType
